1
1

+++ MrYangServer/MediaTools/DirActions/PicCompress.py
@@ -17,6 +17,14 @@ thum = yutils.transform_path(cd_count, yutils.static_media_root, '/pic/thum')
 gif_pic = ''.join([cd_count, yutils.static_root, '/pic/gif_bannder.png'])  # media_source +
 
 
+def middle_out_path(source_path):
+    exten = os.path.splitext(source_path)[1]
+    simple_path = source_path[len(src):]
+    desc_path = middle + '/' + yutils.md5_of_str(os.path.dirname(simple_path))
+    rename_path = desc_path + '/' + yutils.get_md5(source_path) + exten
+    return (rename_path, desc_path)
+
+
 def src2pc(delete_exist):
     ImageFile.LOAD_TRUNCATED_IMAGES = True
     middle_area = 1500 * 1500
@@ -24,14 +32,17 @@ def src2pc(delete_exist):
     # os.system(cmd)
     for root, dirs, files in os.walk(src):
         for file in files:
+            source_path = os.path.join(root, file).replace('\\', '/')
+            (rename_path, desc_path) = middle_out_path(source_path)
+            if yutils.is_gif(file):
+                shutil.copy(source_path, rename_path)
+                continue
             if not yutils.is_photo(file):
                 continue
 
-            source_path = os.path.join(root, file).replace('\\', '/')
-            exten = os.path.splitext(source_path)[1]
-            simple_path = source_path[len(src):]
-            desc_path = middle + '/' + yutils.md5_of_str(os.path.dirname(simple_path))
-            rename_path = desc_path + '/' + yutils.get_md5(source_path) + exten
+            # exten = os.path.splitext(source_path)[1]
+            # desc_path = middle + '/' + yutils.md5_of_str(os.path.dirname(simple_path))
+            # rename_path = desc_path + '/' + yutils.get_md5(source_path) + exten
             if (not delete_exist) and os.path.exists(rename_path):
                 print('文件已存在!' + rename_path)
                 continue
@@ -39,10 +50,7 @@ def src2pc(delete_exist):
                 os.makedirs(desc_path)
             print('源：' + source_path)
 
-            if yutils.is_gif(file):
-                shutil.copy(source_path, rename_path)
-                continue
-                # 直接移动
+            # 直接移动
             if not yutils.is_photo(file):
                 continue
 
@@ -93,7 +101,7 @@ def middle2thum(delete_exist):
             dir = os.path.dirname(desc_path)
             if not os.path.exists(dir):
                 os.makedirs(dir)
-            if yutils.is_gif(file):
+            if yutils.is_gif(file) and os.path.exists(gif_pic):
                 # gif_pic
                 shutil.copy(gif_pic, desc_path)
                 pass
@@ -109,7 +117,7 @@ def middle2thum(delete_exist):
                 region = (0, yoff, w, w + yoff)
 
             crop_img = img.crop(region)  # 保存裁切后的图片
-            crop_img.thumbnail((thum_width,thum_width), Image.ANTIALIAS)
+            crop_img.thumbnail((thum_width, thum_width), Image.ANTIALIAS)
             if 'exif' in img.info:
                 exif_dict = piexif.load(crop_img.info["exif"])
                 exif_bytes = piexif.dump(exif_dict)
@@ -140,12 +148,46 @@ def move_info():
                 # print(source_path)
 
 
+def delete_thum():
+    files = os.listdir(thum)
+    for file in files:
+        path = os.path.join(thum, file)
+        if os.path.isdir(path):
+            print(path)
+            shutil.rmtree(path)
+
+
 # 删除多余的middle 和thum
 def delete_not_exist():
-    pass
+    right_map = {}
+    for root, dirs, files in os.walk(src):
+        for file in files:
+            if not yutils.is_photo(file):
+                continue
+            source_path = os.path.join(root, file).replace('\\', '/')
+            (rename_path, _) = middle_out_path(source_path)
+            if (os.path.exists(rename_path)):
+                right_map[rename_path] = source_path
+
+    delete_list = []
+    for root, dirs, files in os.walk(middle):
+        for file in files:
+            if not yutils.is_photo(file):
+                continue
+            source_path = os.path.join(root, file).replace('\\', '/')
+            if source_path in right_map:
+                pass  # print('' + source_path + "  src:" + right_map[source_path])
+            else:
+                delete_list.append(source_path)
+                # print(source_path)
+    for path in delete_list:
+        os.remove(path)
+    yutils.delete_null_dir(middle)
+    delete_thum()
 
 
 if __name__ == '__main__':
+    delete_not_exist()
     src2pc(False)
     middle2thum(False)
     move_info()
